careergoal routes added,verifyUser middleware corrected

+++ backend/controllers/user.controller.js
@@ -4,9 +4,9 @@ const bcrypt = require("bcrypt");
  const updateprofilecontroller = async (req, res, next) => {
   const userId = req.params.id;
 
-  // if (req.user.userId !== userId) {
-  //   return res.status(403).json({ success: false, message: "unauthorized" });
-  // }
+  if (req.user.userId !== userId) {
+    return res.status(403).json({ success: false, message: "unauthorized" });
+  }
 
   if (req.body.password) {
     req.body.password = bcrypt.hashSync(req.body.password, 10);
@@ -40,9 +40,9 @@ const bcrypt = require("bcrypt");
   const userId = req.params.id;
   // console.log(userId);
   // console.log(req.user.userId);
-  // if (req.user.userId !== userId) {
-  //   return res.status(403).json({ success: false, message: "unauthorized" });
-  // }
+  if (req.user.userId !== userId) {
+    return res.status(403).json({ success: false, message: "unauthorized" });
+  }
   try {
     await User.findByIdAndDelete(userId);
     res.clearCookie("access_token");
@@ -57,8 +57,38 @@ const bcrypt = require("bcrypt");
     res.status(200).json({success:true,message:"user logged out"});}
 }
 
+const updateCareerDetails = async (req, res) => {
+  const userId = req.params.id;
+
+  if (req.user.userId !== userId) {
+    return res.status(403).json({ success: false, message: "unauthorized" });
+  }
+  const { careerGoal, level, subLevel, skills } = req.body;
+  // console.log(req.body);
+  try {
+    const user = await User.findById(userId); 
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    user.careerDetails = {
+      careerGoal,
+      level,
+      subLevel,
+      skills,
+    };
+
+    await user.save();
+    res.status(200).json({ message: "Career details updated successfully", careerDetails: user.careerDetails });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ message: "Failed to update career details" });
+  }
+};
+
 module.exports = {
   logoutuserhandler,
   deleteusercontroller,
   updateprofilecontroller,
+  updateCareerDetails
 };
